refactored some files to be more organized
