+++ src/components/Authentication/LoginPage.js
@@ -6,7 +6,7 @@ import {useLazyQuery, useMutation} from '@apollo/react-hooks';
 import Checkbox from '@material-ui/core/Checkbox';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Grid from '@material-ui/core/Grid';
-import {PasswordInput} from '../Form/Fields';
+import {PasswordInput} from '../Form/Field Components/Fields';
 import TextField from '@material-ui/core/TextField';
 import Typography from '@material-ui/core/Typography';
 import InputAdornment from '@material-ui/core/InputAdornment';
