+++ src/components/Authentication/NewAccount.js
@@ -1,32 +1,36 @@
-import React, { useState } from "react";
-import { Link } from "react-router-dom";
+import React, {useState} from "react";
+import {Link} from "react-router-dom";
 import useStyles from "./styles.js";
 import Typography from "@material-ui/core/Typography";
-import { ResponsiveButton } from '../../theme/ThemedComponents/Button/ResponsiveButton';
-import { PHONE_NUMBER_FIELD, stringField } from "../Form/FormFormats";
-import { FORM_ERROR } from "final-form";
+import {ResponsiveButton} from '../../theme/ThemedComponents/Button/ResponsiveButton';
+import {PHONE_NUMBER_FIELD, stringField} from "../Form/FormFormats";
+import {FORM_ERROR} from "final-form";
 import * as Yup from "yup";
 import gql from "graphql-tag";
-import { useApolloClient } from "@apollo/react-hooks";
-import { Form as ReactForm } from "react-final-form";
+import {useApolloClient} from "@apollo/react-hooks";
+import {Form as ReactForm} from "react-final-form";
 import ToggleButton from "@material-ui/lab/ToggleButton";
 import ToggleButtonGroup from "@material-ui/lab/ToggleButtonGroup";
 import Grid from "@material-ui/core/Grid";
-import * as Fields from "../Form/Fields";
+import * as Fields from "../Form/Field Components/Fields";
 import "./LoginPage.scss";
-import { ReactComponent as Ellipse3 } from "./loginImages/ellipse3.svg";
-import { ReactComponent as Ellipse4 } from "./loginImages/ellipse4.svg";
-import { ReactComponent as Picture2 } from "./loginImages/picture2.svg";
+import {ReactComponent as Ellipse3} from "./loginImages/ellipse3.svg";
+import {ReactComponent as Ellipse4} from "./loginImages/ellipse4.svg";
+import {ReactComponent as Picture2} from "./loginImages/picture2.svg";
 import Step from "@material-ui/core/Step";
 import StepLabel from "@material-ui/core/StepLabel";
 import Stepper from "@material-ui/core/Stepper";
-import { makeValidate } from "mui-rff";
+import {makeValidate} from "mui-rff";
 
 const basicInfo = [
-    <Fields.TextField className="TextField" fullWidth={false} key="firstName" label="First Name" margin="dense" name="firstName" required variant="outlined" />,
-    <Fields.TextField className="TextField" fullWidth={false} key="lastName" label="Last Name" margin="dense" name="lastName" required variant="outlined" />,
-    <Fields.TextField className="TextField" fullWidth={false} key="email" label="Email" margin="dense" name="email" required type="email" variant="outlined" />,
-    <Fields.PasswordInput className="TextField" fullWidth={false} key="password" margin="dense" name="password" required variant="outlined" />,
+    <Fields.TextField className="TextField" fullWidth={false} key="firstName" label="First Name" margin="dense"
+                      name="firstName" required variant="outlined"/>,
+    <Fields.TextField className="TextField" fullWidth={false} key="lastName" label="Last Name" margin="dense"
+                      name="lastName" required variant="outlined"/>,
+    <Fields.TextField className="TextField" fullWidth={false} key="email" label="Email" margin="dense" name="email"
+                      required type="email" variant="outlined"/>,
+    <Fields.PasswordInput className="TextField" fullWidth={false} key="password" margin="dense" name="password" required
+                          variant="outlined"/>,
     <Fields.TextField className="TextField" fullWidth={false} key="phone" label="Phone Number" margin="dense" name="phoneNumber" variant="outlined" />,
 
 ];
