+++ src/components/Form/Field Components/Fields.js
@@ -12,7 +12,7 @@ import * as Fields from "mui-rff";
 import {makeStyles} from "@material-ui/core/styles";
 import {useQuery} from "@apollo/react-hooks";
 import gql from "graphql-tag";
-import {fullName} from "../../utils";
+import {fullName} from "../../../utils";
 import MomentUtils from "@date-io/moment";
 import MaskedInput from "react-text-mask";
 import {Schedule} from "@material-ui/icons";
