+++ src/components/Form/Field Components/TutoringPriceQuote.js
@@ -5,14 +5,14 @@ import PropTypes from "prop-types";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography/Typography";
 // Local Component Imports
-import "./Form.scss";
+import "../Form.scss";
 import TextField from "@material-ui/core/TextField/TextField";
 import MenuItem from "@material-ui/core/MenuItem";
 import InputLabel from "@material-ui/core/InputLabel";
-import {BootstrapInput} from "../FeatureViews/Scheduler/SchedulerUtils";
+import {BootstrapInput} from "../../FeatureViews/Scheduler/SchedulerUtils";
 import gql from "graphql-tag";
 import {useQuery} from "@apollo/react-hooks";
-import Loading from "../OmouComponents/Loading";
+import Loading from "../../OmouComponents/Loading";
 import * as Fields from "mui-rff";
 import {OnChange} from "react-final-form-listeners";
 import makeStyles from "@material-ui/core/styles/makeStyles";
