+++ src/components/Form/OldForm.js
@@ -38,9 +38,8 @@ import VisibilityOff from "@material-ui/icons/VisibilityOff";
 import InputAdornment from "@material-ui/core/InputAdornment";
 // Outside React Component
 import SearchSelect from "react-select";
-import BackButton from "../BackButton.js";
 import Modal from "@material-ui/core/Modal";
-import CompleteCourseRegistration from "./CompleteCourseRegistration";
+import CompleteCourseRegistration from "./Receipts/CompleteCourseRegistration";
 
 import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
@@ -49,7 +48,7 @@ import DialogContentText from "@material-ui/core/DialogContentText";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import {DatePicker, TimePicker} from "@material-ui/pickers";
 import * as utils from "./FormUtils";
-import TutoringPriceQuote from "./TutoringPriceQuote";
+import TutoringPriceQuote from "./Field Components/TutoringPriceQuote";
 import InstructorConflictCheck from "components/InstructorConflictCheck";
 import {combineDateAndTime, durationStringToNum} from "utils";
